update memory layout article
I have a feeling this article is maybe too terse
but I added a couple of example showing overlapping
views and a comment amount `map`.

Maybe I should re-write this article?

I think originally I was going to have a separate typedarray
article but I feel like the 2 topics kind of go hand-in-hand,
or at least as written.