Rename RollingAverage to NonNegativeRollingAverage
And, make it discard negative values.

The spec says timestamp query can return a beginning timestamp
with a value greater than an ending timestamp and such values
should be discarded.

The easiest place to do this is in the RollingAverage class but
just doing it there without renaming it seemed bad since someone
might use RollingAverage for other purposes and not realize it
was discarding negative values.

Further, it's possible offsite samples are linking directly to
rolling-timing.js (which is the case if you export any sample)
so leave that implementation as is and put NonNegativeRollingAverage
in its own file.