Make the HSL diagram start facing the user
I'm not sure what to do here. I wanted to add H,S,L
readings like the RGB diagram has but I couldn't figure
out where to put them that wasn't mis-leading.

Further, it would be nice if the slice of the cylinder
always faced the user? Or maybe it's just a bad diagram.

In any case, people might not interact with them much
and so MVP is just make it face the user to start.